Code was divided into modules
